добавила WriteTrainer()
